requirements.txt
python-telegram-bot[job-queue]>=20.0
APScheduler>=3.6.3